Update on sign up feature, fixed header issue and created signout option
header.html now works to hold the menu bar so now to get the header in a html file, you just need to call it.
Checked the length of user password, which now has to be 7 characters or more.
Updated the email error checking to make sure that it was a valid email format being entered.
User can now sign out of the site and will be redirected back to the register.html page.